Inital commit for adding a direct package reference member to script objects.
