+++ pol-core/bscript/objmembers.h
@@ -285,6 +285,7 @@ enum MemberID
   MBR_PHYSICAL_RESIST_CAP_MOD,  // 245
   MBR_LUCK_MOD,
   MBR_HIT_CHANCE_MOD,
+  MBR_PACKAGE,
 };
