+++ pol-core/pol/exscrobj.cpp
@@ -27,6 +27,7 @@
 #include "../clib/strutil.h"
 #include "../plib/systemstate.h"
 #include "module/uomod.h"
+#include "module/polsystemmod.h"
 #include "polcfg.h"
 #include "uoexec.h"
 #include "uoscrobj.h"
@@ -176,6 +177,8 @@ BObjectRef ScriptExObjImp::get_member_id( const int id )
     return BObjectRef( new BLong( uoexec->pid() ) );
   case MBR_NAME:
     return BObjectRef( new String( uoexec->scriptname() ) );
+  case MBR_PACKAGE:
+    return BObjectRef( new Pol::Module::PackageObjImp( uoexec->prog()->pkg ) );
   case MBR_STATE:
     return BObjectRef( new String( uoexec->state() ) );
   case MBR_INSTR_CYCLES:
